Implemented direct rendering support.
This allows most comms to go via the (more full-featured) cloud service
API, but rendering requests are sent directly to the hub. This can
result in significant time savings (25s vs 15s for a test render) as
well as much faster latency between sending the render request and the
first sign of a change on the display. This is especially useful for
displays with buttons, as it lessens the feeling that nothing happened
when pushing a button.